Move util functions to reports.py and logging.py
