+++ movement/filtering.py
@@ -1,109 +1,10 @@
 """Filter and interpolate  pose tracks in ``movement`` datasets."""
 
-import logging
-from datetime import datetime
-from functools import wraps
-
 import xarray as xr
 from scipy import signal
 
-from movement.utils.logging import log_error
-
-
-def log_to_attrs(func):
-    """Log the operation performed by the wrapped function.
-
-    This decorator appends log entries to the data's ``log``
-    attribute. The wrapped function must accept an ``xarray.Dataset``
-    or ``xarray.DataArray`` as its first argument and return an
-    ``xarray.Dataset`` or ``xarray.DataArray``.
-    """
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        result = func(*args, **kwargs)
-
-        log_entry = {
-            "operation": func.__name__,
-            "datetime": str(datetime.now()),
-            **{f"arg_{i}": arg for i, arg in enumerate(args[1:], start=1)},
-            **kwargs,
-        }
-
-        # Append the log entry to the result's attributes
-        if result is not None and hasattr(result, "attrs"):
-            if "log" not in result.attrs:
-                result.attrs["log"] = []
-            result.attrs["log"].append(log_entry)
-
-        return result
-
-    return wrapper
-
-
-def generate_nan_report(
-    data: xr.DataArray, keypoint: str, individual: str | None = None
-) -> str:
-    """Generate NaN value report for a given keypoint and individual.
-
-    This function calculates the number and percentage of NaN points
-    for a given keypoint and individual in the input data. A keypoint
-    is considered NaN if any of its ``space`` coordinates are NaN.
-
-    Parameters
-    ----------
-    data : xarray.DataArray
-        The input data containing ``keypoints`` and ``individuals``
-        dimensions.
-    keypoint : str
-        The name of the keypoint for which to generate the report.
-    individual : str, optional
-        The name of the individual for which to generate the report.
-
-    Returns
-    -------
-    str
-        A string containing the report.
-
-    """
-    selectedta = (
-        data.sel(individuals=individual, keypoints=keypoint)
-        if individual
-        else data.sel(keypoints=keypoint)
-    )
-    n_nans = selectedta.isnull().any(["space"]).sum(["time"]).item()
-    n_points = selectedta.time.size
-    percent_nans = round((n_nans / n_points) * 100, 1)
-    return f"\n\t\t{keypoint}: {n_nans}/{n_points} ({percent_nans}%)"
-
-
-def report_nan_values(da: xr.DataArray, label: str | None = None):
-    """Report the number and percentage of keypoints that are NaN.
-
-    Numbers are reported for each individual and keypoint in the dataset.
-
-    Parameters
-    ----------
-    da : xarray.DataArray
-        The input data containing pose tracks and metadata.
-    label : str, optional
-        Label to identify the dataset in the report. If not provided,
-        the name of the DataArray is used as the label. Default is None.
-
-    """
-    # Compile the report
-    if not label:
-        label = da.name
-    nan_report = f"\nMissing points (marked as NaN) in {label}"
-    for ind in da.individuals.values:
-        nan_report += f"\n\tIndividual: {ind}"
-        for kp in da.keypoints.values:
-            nan_report += generate_nan_report(da, kp, individual=ind)
-    # Write nan report to logger
-    logger = logging.getLogger(__name__)
-    logger.info(nan_report)
-    # Also print the report to the console
-    print(nan_report)
+from movement.utils.logging import log_error, log_to_attrs
+from movement.utils.reports import report_nan_values
 
 
 @log_to_attrs
